Check state more often when radio is playing music
The stream seems to be a bit behind actual radio transmission, but the
now playing data matches more closely to actual radio. This change
tries to change to hosts as soon as possible so that the beginning
of a talk is not lost. Also slower checks when hosts are speaking
hopefully results in less cutting the final words before music.